Revert to working version

+++ src/main/java/fi/haagahelia/wiki_data_crud/service/DropTableService.java
@@ -28,12 +28,16 @@ public class DropTableService {
         return dropTableRepository.save(dropTable);
     }
 
-    public Optional<DropTable> updateDropTable(Long id, DropTable dropTableDetails) {
-        return dropTableRepository.findById(id).map(dropTable -> {
+    public DropTable updateDropTable(Long id, DropTable dropTableDetails) {
+        Optional<DropTable> optionalDropTable = dropTableRepository.findById(id);
+        if (optionalDropTable.isPresent()) {
+            DropTable dropTable = optionalDropTable.get();
             dropTable.setMonster(dropTableDetails.getMonster());
             dropTable.setItems(dropTableDetails.getItems());
             return dropTableRepository.save(dropTable);
-        });
+        } else {
+            return null;
+        }
     }
 
     public boolean deleteDropTable(Long id) {
